Removed Metapi from class name
Summary:
The library and package already qualifies that this is
MET API utils, so no need to call the class by the same
name.

Also, the old version was capitalized wrong (MetApi, not Metapi).

Worth fixing now, before we are using this in multiple applications.

This will break the build. I'll push a fix for metapi-observations
ASAP.

Test Plan: `activator clean test`

Reviewers: #met_api

Differential Revision: https://phab.met.no/D730